\name{demogRse-package}
\alias{demogRse-package}
\alias{demogRse}
\docType{package}
\title{
To patch Tuljapurkar et al.'s elasticities of lambdaS into demogR
}
\description{
To patch Tuljapurkar et al.'s elasticities of lambdaS into demogR; generate Markov chains to describe random environments, project populations through these environments and calculate multiple elasticities of long-run stochastic growth (lambdaS).  These functions are based upon MATLAB code developed by Horvitz, Haridas and Tuljapurkar during that period.
}
\details{

\tabular{ll}{
Package: \tab demogRse\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-06-08\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
Thomas H.G. Ezard, Carol C. Horvitz, C.V. Haridas and S. Tuljapurkar.

Correspondence: Thomas Ezard <thomas.ezard@imperial.ac.uk>
}

\references{
If you use these functions, it'd be really great if you'd cite either:

Ezard, T. H. G., J.-M. Gaillard, M. J. Crawley, and T. Coulson. 2008. Habitat dependence and correlations between elasticities of long-term growth rates. American Naturalist, 172: 424-430.

or

Ezard, T. H. G., and T. Coulson. in resubmission. How sensitive are elasticities of long-run stochastic growth to how environmental variability is modelled? 


The theory was developed in the following manuscripts:

Tuljapurkar, S., Horvitz, C. C., and J. B. Pascarella (2003) The many growth rates and elasticities of populations in random environments. Am. Nat. 162: 489-502.

Haridas, C. V., and S. Tuljapurkar. (2005) Elasticities in variable environments: properties and implications. Am. Nat. 166: 481-495.

Horvitz, C. C., S. Tuljapurkar, and J. B. Pascarella. 2005. Plant-animal interactions in random environments: habitat stage elasticity, seed predators, and hurricanes. Ecology 86:3312-3322.

Morris, W. F., S. Tuljapurkar, C. V. Haridas, E. S. Menges, C. C. Horvitz, and C. A. Pfister. 2006. Sensitivity of the population growth rate to demographic variability within and between phases of the disturbance cycle. Ecology Letters 9:1331-1341.

}

\seealso{\code{\link{stochPerturb}}}
\keyword{ package }
